update ver 0.1.6

+++ insight_android_core/Insight/src/main/java/kr/co/bizspring/insight/lib/tracker/Profiler.java
@@ -85,7 +85,7 @@ public class Profiler {
         return profileDB.getInt(StaticValues.PARAM_LTRVNC);
     }
     public long getLtrvni() {
-        return profileDB.getInt(StaticValues.PARAM_LTRVNI);
+        return profileDB.getLong(StaticValues.PARAM_LTRVNI);
     }
 
     public void notifyLtrvnc(){
